refactor: domain package 구조 개선 (#20)
